requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
falcon==1.4.1
idna==2.7
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.8.0
python-mimeparse==1.6.0
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.14
sqlservice==1.1.3
stripe==2.12.0
urllib3==1.24.1
zipcodes==1.0.4
